+++ app/models/account/category.rb
@@ -1,0 +1,9 @@
+module Account
+  class Category < ApplicationRecord
+    # ASSOCIATIONS
+    belongs_to :user
+
+    # VALIDATIONS
+    validates :name, presence: true, uniqueness: true
+  end
+end
