--- app/models/category.rb
@@ -1,7 +1,0 @@
-class Category < ApplicationRecord
-  # ASSOCIATIONS
-  belongs_to :user
-
-  # VALIDATIONS
-  validates :name, presence: true, uniqueness: true
-end
